requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.100.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
openai>=1.0.0
